Fix phrasing of UG for filter and find feature
Previously, the phrasing of the UG caused the reader to have a
misconception in testing.

What we meant in the UG was the opposite of what the tester had though
of it as. Example: if there is a tenant with address in the app that has
"Kent", if the user type "filter Ken", the tenant will appear. What the
tester thought is that if the user type 'filter Kents', the tenant with
address "Kent" should appear.

Fixes #128