Fix Bug in DG whereby manual testing was unclear
Previously, readers had trouble identifying what the test case is meant
to test as it was vague regarding whether one of the tenants can have
"12345678" as a phone number.

Now, the commands 'add' 'paid' and 'unpaid' in the
'Appendix: Instructions for manual testing' specifies that the test case
fails because '12345678' is an invalid number.

This update ensures that readers are informed that the test cases with
'12345678' fails because '12345678' is an invalid number.

Fixes #151

Additionally, fix bug whereby the User Stories had wrong priorities
given to them.