Fix mismatch in filter feature specification between UG and DG
Previously, in the DG, the Filter Command is specified as being able to
filter tenants based on multiple fields such as address. However,
in the User Guide, the Filter Command is specified as only being able
to filter the list for tenants whose address contains any of the given
keywords, where Address is the only filterable field.

Now, the Filter Command in the DG is specified as being able to filter
tenants only based on their addresses.

Fixes #154

Edit the paid and unpaid feature description under the
'Newly Added Features' in the DG.