Add Paid and Unpaid features to HelpWindow
Previously, the Help Window did not contain the Unpaid and Paid command.
Adding the Paid and Unpaid Command to the HelpWindow ensures users are
aware of these 2 commands also.

Additionally, update the paid details in the UG and DG to reflect
valid Singaporean phone number and error message display of the paid
and unpaid command.