Add Screenshot examples to the UG for the different features
Screenshots were only added to some features to avoid cluttering the UG
with too many screenshots. Fro example, more intuitive commands like
unpaid or unarchive which are the opposite of paid and archive do not
have screenshot examples.

Additionally, ensure that the screenshot sizes and frame, along with the
tenant details are standardized and consistent